==> 8. Use Cases/anomaly-detection-with-augmentation/2_profile_data.ipynb <==
# Profile Data 

Data Profiling through YData's Pandas Profiling. 

## 0 - Setup 

### 0.1 - Package

%%capture 
!pip install pandas-profiling

### 0.2 - Imports 

import pandas as pd
from  pandas_profiling import ProfileReport

from ydata.connectors import LocalConnector
from ydata.connectors.filetype import FileType
from ydata.utils.formats import read_json

## 1 - Load Data

# Read data
df = pd.read_csv('data.csv')

## 2 - Profile Data

profile = ProfileReport(df, title="Network Mobile")

## 3 - Create Artifacts 

import json
 # Artifact 
metadata = {
 'outputs' : [{
   'type': 'web-app',
   'storage': 'inline',
   'source': profile.to_html(),
 }]
}

with open('mlpipeline-ui-metadata.json', 'w') as metadata_file:
    json.dump(metadata, metadata_file)
