==> 8. Use Cases/anomaly-detection-with-augmentation/3_preprocess_data.ipynb <==
# Preprocess Data

# 0 - Setup

### 0.1 Packages 

import pandas as pd
from ydata.connectors import LocalConnector
from ydata.connectors.filetype import FileType
from ydata.utils.formats import read_json

## 1 - Load Data

# Read the data 
data = pd.read_csv('data.csv')

## 2 - Data Processing 

# Do the little preprocessing
data.drop_duplicates(inplace=True)
data.drop(['meanUE_DL', 'maxUE_DL', 'maxUE_UL'], axis=1, inplace=True)

## 3 - Store Data 

# Output the changed data

data.to_csv('data_processed.csv', index=False)
