==> 8. Use Cases/anomaly-detection-with-augmentation/5_synthesize_data.ipynb <==
# Synthesize Data 

## 0 - Setup

### 0.1 - Packages 

import pandas as pd
from numpy import repeat

from ydata.metadata import Metadata
from ydata.synthesizers.regular import RegularSynthesizer
from ydata.connectors import LocalConnector
from ydata.connectors.filetype import FileType
from ydata.dataset import Dataset

## 1 - Load Data

# Read the data (Only if you want to train a new network)
data = pd.read_csv('data_processed.csv')
#Filter the data by the unusual observations
data = data.query("Unusual == 1")

data = data.drop(['Time', 'CellName'], axis=1)

## 2 - Synthesization 

### 2.1 - Train the Synthesizer 

#create the dataset. This allows scale and integration with YData's platform SDK's
dataset = Dataset(data)
#Calculate the metadata
m = Metadata(dataset)

print(m)

# Initialize and Load an alerady trained Synthesizer.
network_synth = RegularSynthesizer()

network_synth.fit(dataset, metadata=m)
network_synth.save('trained_model.pkl')

#Load a pre-saved model
#network_synth = network_synth.load('trained_model.pkl')

### 2.2 - Sample Data 

# Create the samples and flag them.
synth_sample = network_synth.sample(5000).to_pandas()

# Flag the created samples and assign columns 
synth_sample.index = repeat(999999, len(synth_sample))
synth_sample.columns = data.columns

### 3 - Store Data 

synth_sample.to_csv('data_sampled.csv', index=False)
